Create file-saver service
